Fix logic check to support config tab suppled build mode value Make Step order value type be the same as TestStepDTO (for reporting service) order value type.

+++ codeless_core/src/main/java/com/tmobile/ct/codeless/core/Step.java
@@ -111,12 +111,12 @@ public interface Step extends Executable, Validatable{
 	 *
 	 * @param order
 	 */
-	void setOrder (Long order);
+	void setOrder (Integer order);
 
 	/**
 	 * Gets the step order.
 	 *
 	 * @return the order
 	 */
-	Long getOrder ();
+	Integer getOrder ();
 }
